# api/openapi.yaml
openapi: 3.0.3
info:
  title: Inventory Management API
  version: 0.4.0
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
  /api/v1/auth/login:
    post:
      summary: Login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Invalid credentials / account disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /api/v1/auth/register:
    post:
      summary: Register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: Register result (also logs in)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /api/v1/auth/refresh:
    post:
      summary: Refresh access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: New tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Invalid/expired refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /api/v1/auth/change-password:
    post:
      summary: Change password (revokes refresh tokens)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        storeId:
          type: integer
          format: int64
          nullable: true
      required: [email, password]
    RefreshTokenRequest:
      type: object
      properties:
        refreshToken:
          type: string
      required: [refreshToken]
    ChangePasswordRequest:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
      required: [currentPassword, newPassword]
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresInSeconds:
          type: integer
          format: int64
        user:
          $ref: "#/components/schemas/UserResponse"
      required: [accessToken, refreshToken, expiresInSeconds, user]
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          format: email
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        role:
          type: string
          enum: [EMPLOYEE, MANAGER]
        storeId:
          type: integer
          format: int64
          nullable: true
        storeName:
          type: string
          nullable: true
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, email, role, active, createdAt]
    ApiError:
      type: object
      properties:
        message:
          type: string
        timestamp:
          type: string
          format: date-time
      required: [message, timestamp]
...